+++ CHANGELOG.rst
@@ -1,1 +1,2 @@
-1.0: HTTP client & server
+* 1.1: Can start server at random open port, very useful for tests
+* 1.0: HTTP client & server